requirements.txt
fastapi
uvicorn
gradio
requests
openai
python-dotenv
google-generativeai
